Moving AzureOpenAIAPIType enum
