+++ src/python/PythonSDK/foundationallm/langchain/language_models/openai/__init__.py
@@ -1,5 +1,4 @@
 """OpenAI language model module"""
-from .azure_openai_api_types import AzureOpenAIAPIType
 from .openai_model_base import OpenAIModelBase
 from .azure_chat_model import AzureChatModel
 from .azure_text_completion_model import AzureTextCompletionModel
