+++ src/python/PythonSDK/foundationallm/langchain/language_models/openai/azure_chat_model.py
@@ -2,8 +2,8 @@ from langchain.base_language import BaseLanguageModel
 from langchain.chat_models import AzureChatOpenAI
 
 from foundationallm.config import Configuration
-from foundationallm.langchain.language_models.openai import OpenAIModelBase, AzureOpenAIAPIType
-from foundationallm.models.language_models import LanguageModel
+from foundationallm.langchain.language_models.openai import OpenAIModelBase
+from foundationallm.models.language_models import LanguageModel, AzureOpenAIAPIType
 
 class AzureChatModel(OpenAIModelBase):
     """Azure OpenAI chat model."""
