+++ src/python/PythonSDK/foundationallm/models/language_models/__init__.py
@@ -1,4 +1,5 @@
 """Language model models module"""
+from .azure_openai_api_types import AzureOpenAIAPIType
 from .language_model_provider import LanguageModelProvider
 from .language_model_type import LanguageModelType
 from .language_model import LanguageModel
