+++ src/python/PythonSDK/foundationallm/models/language_models/azure_openai_api_types.py
